@@ -3,20 +3,29 @@ package provider
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
+	"database/sql"
+	"encoding/base64"
+	"encoding/hex"
+	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"math/rand"
 	"net/http"
 	"net/url"
+	"regexp"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
+	"price-feeder/oracle/provider/volume"
 	"price-feeder/oracle/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/rs/zerolog"
+	"golang.org/x/crypto/sha3"
 )
 
 const (
@@ -24,44 +33,62 @@ const (
 	staleTickersCutoff   = 1 * time.Minute
 	providerCandlePeriod = 10 * time.Minute
 
-	ProviderAstroportTerra2    Name = "astroport_terra2"
-	ProviderAstroportNeutron   Name = "astroport_neutron"
 	ProviderAstroportInjective Name = "astroport_injective"
-	ProviderFin                Name = "fin"
-	ProviderFinV2              Name = "finv2"
-	ProviderKraken             Name = "kraken"
+	ProviderAstroportNeutron   Name = "astroport_neutron"
+	ProviderAstroportTerra2    Name = "astroport_terra2"
 	ProviderBinance            Name = "binance"
 	ProviderBinanceUS          Name = "binanceus"
-	ProviderCamelotV2          Name = "camelotv2"
-	ProviderCamelotV3          Name = "camelotv3"
-	ProviderOsmosis            Name = "osmosis"
-	ProviderOsmosisV2          Name = "osmosisv2"
-	ProviderHuobi              Name = "huobi"
-	ProviderOkx                Name = "okx"
-	ProviderGate               Name = "gate"
-	ProviderCoinbase           Name = "coinbase"
-	ProviderBitget             Name = "bitget"
-	ProviderBitmart            Name = "bitmart"
-	ProviderBkex               Name = "bkex"
+	ProviderBingx              Name = "bingx"
 	ProviderBitfinex           Name = "bitfinex"
 	ProviderBitforex           Name = "bitforex"
+	ProviderBitget             Name = "bitget"
+	ProviderBitmart            Name = "bitmart"
 	ProviderBitstamp           Name = "bitstamp"
-	ProviderHitBtc             Name = "hitbtc"
-	ProviderPoloniex           Name = "poloniex"
-	ProviderPyth               Name = "pyth"
-	ProviderPhemex             Name = "phemex"
-	ProviderLbank              Name = "lbank"
-	ProviderKucoin             Name = "kucoin"
+	ProviderBkex               Name = "bkex"
 	ProviderBybit              Name = "bybit"
-	ProviderMexc               Name = "mexc"
+	ProviderCamelotV2          Name = "camelotv2"
+	ProviderCamelotV3          Name = "camelotv3"
+	ProviderCoinbase           Name = "coinbase"
+	ProviderCoinex             Name = "coinex"
 	ProviderCrypto             Name = "crypto"
 	ProviderCurve              Name = "curve"
-	ProviderMock               Name = "mock"
-	ProviderStride             Name = "stride"
-	ProviderXt                 Name = "xt"
+	ProviderDexter             Name = "dexter"
+	ProviderFin                Name = "fin"
+	ProviderFinV2              Name = "finv2"
+	ProviderGate               Name = "gate"
+	ProviderHelix              Name = "helix"
+	ProviderHitBtc             Name = "hitbtc"
+	ProviderHuobi              Name = "huobi"
 	ProviderIdxOsmosis         Name = "idxosmosis"
-	ProviderZero               Name = "zero"
+	ProviderKraken             Name = "kraken"
+	ProviderKucoin             Name = "kucoin"
+	ProviderLbank              Name = "lbank"
+	ProviderMaya               Name = "maya"
+	ProviderMexc               Name = "mexc"
+	ProviderMock               Name = "mock"
+	ProviderOkx                Name = "okx"
+	ProviderOsmosis            Name = "osmosis"
+	ProviderOsmosisV2          Name = "osmosisv2"
+	ProviderPancakeV3Bsc       Name = "pancakev3_bsc"
+	ProviderPhemex             Name = "phemex"
+	ProviderPionex             Name = "pionex"
+	ProviderPoloniex           Name = "poloniex"
+	ProviderPyth               Name = "pyth"
+	ProviderShade              Name = "shade"
+	ProviderStride             Name = "stride"
 	ProviderUniswapV3          Name = "uniswapv3"
+	ProviderUnstake            Name = "unstake"
+	ProviderVelodromeV2        Name = "velodromev2"
+	ProviderWhitewhaleCmdx     Name = "whitewhale_cmdx"
+	ProviderWhitewhaleHuahua   Name = "whitewhale_huahua"
+	ProviderWhitewhaleInj      Name = "whitewhale_inj"
+	ProviderWhitewhaleJuno     Name = "whitewhale_juno"
+	ProviderWhitewhaleLuna     Name = "whitewhale_luna"
+	ProviderWhitewhaleLunc     Name = "whitewhale_lunc"
+	ProviderWhitewhaleSei      Name = "whitewhale_sei"
+	ProviderWhitewhaleWhale    Name = "whitewhale_whale"
+	ProviderXt                 Name = "xt"
+	ProviderZero               Name = "zero"
 )
 
 type (
@@ -83,6 +110,7 @@ type (
 
 	provider struct {
 		ctx       context.Context
+		name      string
 		endpoints Endpoint
 		httpBase  string
 		http      *http.Client
@@ -93,6 +121,10 @@ type (
 		tickers   map[string]types.TickerPrice
 		contracts map[string]string
 		websocket *WebsocketController
+		db        *sql.DB
+		volumes   volume.VolumeHandler
+		height    uint64
+		chain     string
 	}
 
 	PollingProvider interface {
@@ -120,6 +152,22 @@ type (
 		PingType          uint
 		PingMessage       string
 		ContractAddresses map[string]string
+		VolumeBlocks      int
+		VolumePause       int
+		Decimals          map[string]int
+		Periods           map[string]int
+	}
+
+	EvmLog struct {
+		Address string   `json:"address"`
+		Topics  []string `json:"topics"`
+		Data    string   `json:"data"`
+		Number  string   `json:"blockNumber"`
+		Height  uint64
+	}
+
+	EvmBlock struct {
+		Timestamp string `json:"timestamp"`
 	}
 )
 
@@ -134,9 +182,20 @@ func (p *provider) Init(
 	p.ctx = ctx
 	p.endpoints = endpoints
 	p.endpoints.SetDefaults()
+
+	// remove trailing slashes
+	for i, url := range p.endpoints.Urls {
+		p.endpoints.Urls[i] = strings.TrimRight(url, "/")
+	}
+
 	p.logger = logger.With().Str("provider", p.endpoints.Name.String()).Logger()
 	p.tickers = map[string]types.TickerPrice{}
 	p.http = newDefaultHTTPClient()
+
+	if len(p.endpoints.Urls) == 0 {
+		p.logger.Error().Msg("no endpoint urls found")
+		return
+	}
 	p.httpBase = p.endpoints.Urls[0]
 
 	p.contracts = endpoints.ContractAddresses
@@ -161,6 +220,52 @@ func (p *provider) Init(
 		)
 		go p.websocket.Start()
 	}
+
+	// set contract<>symbol mapping
+
+	p.contracts = endpoints.ContractAddresses
+
+	for symbol, contract := range endpoints.ContractAddresses {
+		p.contracts[contract] = symbol
+	}
+
+	p.height = 0
+
+	if p.db == nil {
+		return
+	}
+
+	// set up volume handler
+
+	symbols := []string{}
+	for _, pair := range pairs {
+		skip := false
+		for _, symbol := range []string{pair.Base, pair.Quote} {
+			_, found := endpoints.Decimals[symbol]
+			if !found {
+				skip = true
+				logger.Debug().
+					Str("symbol", symbol).
+					Msg("unknown decimal")
+			}
+		}
+		if skip {
+			continue
+		}
+
+		symbols = append(symbols, pair.Base+pair.Quote)
+		symbols = append(symbols, pair.Quote+pair.Base)
+	}
+
+	var period int64 = 86400
+	name := endpoints.Name.String()
+
+	volumes, err := volume.NewVolumeHandler(logger, p.db, name, symbols, period)
+	if err != nil {
+		panic(err)
+	}
+
+	p.volumes = volumes
 }
 
 func (p *provider) GetTickerPrices(pairs ...types.CurrencyPair) (map[string]types.TickerPrice, error) {
@@ -217,6 +322,231 @@ func (p *provider) CurrencyPairToProviderPair(pair types.CurrencyPair) string {
 	return pair.String()
 }
 
+func (p *provider) compactJsonString(message string) (string, error) {
+	buffer := new(bytes.Buffer)
+	err := json.Compact(buffer, []byte(message))
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return "", err
+	}
+
+	return buffer.String(), nil
+}
+
+func (p *provider) getCosmosTx(hash string) (types.CosmosTx, error) {
+	type (
+		Attribute struct {
+			Key   string `json:"key"`
+			Value string `json:"value"`
+		}
+
+		Event struct {
+			Type       string      `json:"type"`
+			Attributes []Attribute `json:"attributes"`
+		}
+
+		TxResponse struct {
+			Code   int64   `json:"code"`
+			Height string  `json:"height"`
+			Time   string  `json:"timestamp"`
+			Events []Event `json:"events"`
+		}
+
+		Response struct {
+			TxResponse TxResponse `json:"tx_response"`
+		}
+	)
+
+	var tx types.CosmosTx
+
+	path := fmt.Sprintf("/cosmos/tx/v1beta1/txs/%s", hash)
+
+	content, err := p.httpGet(path)
+	if err != nil {
+		return tx, err
+	}
+
+	var response Response
+
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		return tx, err
+	}
+
+	timestamp, err := time.Parse(time.RFC3339, response.TxResponse.Time)
+	if err != nil {
+		return tx, err
+	}
+
+	height, err := strconv.ParseUint(response.TxResponse.Height, 10, 64)
+	if err != nil {
+		return tx, err
+	}
+
+	events := make([]types.CosmosTxEvent, len(response.TxResponse.Events))
+	for i, event := range response.TxResponse.Events {
+		events[i] = types.CosmosTxEvent{
+			Type:       event.Type,
+			Attributes: map[string]string{},
+		}
+
+		for _, attribute := range event.Attributes {
+			events[i].Attributes[attribute.Key] = attribute.Value
+		}
+	}
+
+	tx = types.CosmosTx{
+		Code:   response.TxResponse.Code,
+		Time:   timestamp,
+		Height: height,
+		Events: events,
+		Hash:   hash,
+	}
+
+	return tx, nil
+}
+
+func (p *provider) getCosmosHeight() (uint64, error) {
+	path := "/cosmos/base/tendermint/v1beta1/blocks/latest"
+	content, err := p.httpGet(path)
+	if err != nil {
+		return 0, err
+	}
+
+	var response types.CosmosBlockResponse
+
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		return 0, err
+	}
+
+	height, err := strconv.ParseUint(response.Block.Header.Height, 10, 64)
+	if err != nil {
+		return 0, err
+	}
+
+	return height, nil
+}
+
+func (p *provider) getCosmosTxs(
+	height uint64,
+	msgTypes []string,
+) ([]types.CosmosTx, time.Time, error) {
+	var timestamp time.Time
+	if height == 0 {
+		return nil, timestamp, fmt.Errorf("height is 0")
+	}
+
+	var response types.CosmosBlockResponse
+
+	path := fmt.Sprintf("/cosmos/tx/v1beta1/txs/block/%d", height)
+	content, err := p.httpGet(path)
+	if err != nil {
+		return nil, timestamp, err
+	}
+
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		return nil, timestamp, err
+	}
+
+	timestamp, err = time.Parse(time.RFC3339, response.Block.Header.Time)
+	if err != nil {
+		return nil, timestamp, err
+	}
+
+	hashes := []string{}
+	filter := map[string]struct{}{}
+	for _, msgType := range msgTypes {
+		filter[msgType] = struct{}{}
+	}
+
+	for i, tx := range response.Txs {
+		for _, message := range tx.Body.Messages {
+			_, found := filter[message.Type]
+			if !found {
+				continue
+			}
+
+			data, err := base64.StdEncoding.DecodeString(
+				response.Block.Data.Txs[i],
+			)
+			if err != nil {
+				return nil, timestamp, err
+			}
+
+			hash := sha256.New()
+			hash.Write(data)
+			hashes = append(hashes, fmt.Sprintf("%X", hash.Sum(nil)))
+		}
+	}
+
+	txs := []types.CosmosTx{}
+	for _, hash := range hashes {
+		tx, err := p.getCosmosTx(hash)
+		if err != nil {
+			return nil, timestamp, err
+		}
+		txs = append(txs, tx)
+	}
+
+	return txs, timestamp, nil
+}
+
+func (p *provider) wasmRawQuery(contract, message string) ([]byte, error) {
+	bz := append([]byte{0}, []byte(message)...)
+	query := base64.StdEncoding.EncodeToString(bz)
+
+	path := fmt.Sprintf(
+		"/cosmwasm/wasm/v1/contract/%s/raw/%s",
+		contract, query,
+	)
+
+	content, err := p.httpGet(path)
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return nil, err
+	}
+
+	var response struct {
+		Data string `json:"data"`
+	}
+
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		return nil, err
+	}
+
+	data, err := base64.StdEncoding.DecodeString(response.Data)
+	if err != nil {
+		return nil, err
+	}
+
+	return data, nil
+}
+
+func (p *provider) wasmSmartQuery(contract, message string) ([]byte, error) {
+	message, err := p.compactJsonString(message)
+	if err != nil {
+		return nil, err
+	}
+
+	query := base64.StdEncoding.EncodeToString([]byte(message))
+
+	path := fmt.Sprintf(
+		"/cosmwasm/wasm/v1/contract/%s/smart/%s",
+		contract, query,
+	)
+
+	content, err := p.httpGet(path)
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return nil, err
+	}
+
+	return content, nil
+}
+
 func (p *provider) httpGet(path string) ([]byte, error) {
 	return p.httpRequest(path, "GET", nil, nil)
 }
@@ -231,14 +561,24 @@ func (p *provider) httpPost(path string, body []byte) ([]byte, error) {
 func (p *provider) httpRequest(path string, method string, body []byte, headers map[string]string) ([]byte, error) {
 	res, err := p.makeHttpRequest(p.httpBase+path, method, body, headers)
 	if err != nil {
-		p.logger.Warn().
-			Str("endpoint", p.httpBase).
-			Str("path", path).
-			Msg("trying alternate http endpoints")
-		for _, endpoint := range p.endpoints.Urls {
-			if endpoint == p.httpBase {
-				continue
+		index := 0
+		urls := []string{}
+
+		for i, url := range p.endpoints.Urls {
+			if p.httpBase == url {
+				index = i
+				break
 			}
+		}
+
+		urls = append(urls, p.endpoints.Urls[index+1:]...)
+		urls = append(urls, p.endpoints.Urls[:index]...)
+
+		for _, endpoint := range urls {
+			p.logger.Warn().
+				Str("endpoint", endpoint).
+				Msg("trying alternate http endpoints")
+
 			res, err = p.makeHttpRequest(endpoint+path, method, body, headers)
 			if err == nil {
 				p.logger.Info().Str("endpoint", endpoint).Msg("selected alternate http endpoint")
@@ -271,6 +611,9 @@ func (p *provider) makeHttpRequest(url string, method string, body []byte, heade
 	if res.StatusCode != 200 {
 		p.logger.Warn().
 			Int("code", res.StatusCode).
+			Str("body", string(body)).
+			Str("url", url).
+			Str("method", method).
 			Msg("http request returned invalid status")
 		if res.StatusCode == 429 || res.StatusCode == 418 {
 			p.logger.Warn().
@@ -280,9 +623,13 @@ func (p *provider) makeHttpRequest(url string, method string, body []byte, heade
 		}
 		return nil, fmt.Errorf("http request returned invalid status")
 	}
-	content, err := ioutil.ReadAll(res.Body)
+	content, err := io.ReadAll(res.Body)
 	if err != nil {
 		return nil, err
+	}
+
+	if len(content) == 0 {
+		return nil, fmt.Errorf("empty response")
 	}
 	return content, nil
 }
@@ -302,6 +649,8 @@ func (e *Endpoint) SetDefaults() {
 		defaults = bitfinexDefaultEndpoints
 	case ProviderBinanceUS:
 		defaults = binanceUSDefaultEndpoints
+	case ProviderBingx:
+		defaults = bingxDefaultEndpoints
 	case ProviderBitget:
 		defaults = bitgetDefaultEndpoints
 	case ProviderBitmart:
@@ -318,16 +667,22 @@ func (e *Endpoint) SetDefaults() {
 		defaults = camelotV3DefaultEndpoints
 	case ProviderCoinbase:
 		defaults = coinbaseDefaultEndpoints
+	case ProviderCoinex:
+		defaults = coinexDefaultEndpoints
 	case ProviderCrypto:
 		defaults = cryptoDefaultEndpoints
 	case ProviderCurve:
 		defaults = curveDefaultEndpoints
+	case ProviderDexter:
+		defaults = dexterDefaultEndpoints
 	case ProviderFin:
 		defaults = finDefaultEndpoints
 	case ProviderFinV2:
 		defaults = finV2DefaultEndpoints
 	case ProviderGate:
 		defaults = gateDefaultEndpoints
+	case ProviderHelix:
+		defaults = helixDefaultEndpoints
 	case ProviderHitBtc:
 		defaults = hitbtcDefaultEndpoints
 	case ProviderIdxOsmosis:
@@ -342,6 +697,8 @@ func (e *Endpoint) SetDefaults() {
 		defaults = lbankDefaultEndpoints
 	case ProviderMexc:
 		defaults = mexcDefaultEndpoints
+	case ProviderMaya:
+		defaults = mayaDefaultEndpoints
 	case ProviderMock:
 		defaults = mockDefaultEndpoints
 	case ProviderOkx:
@@ -350,14 +707,40 @@ func (e *Endpoint) SetDefaults() {
 		defaults = osmosisDefaultEndpoints
 	case ProviderOsmosisV2:
 		defaults = osmosisv2DefaultEndpoints
+	case ProviderPancakeV3Bsc:
+		defaults = PancakeV3BscDefaultEndpoints
 	case ProviderPhemex:
 		defaults = phemexDefaultEndpoints
+	case ProviderPionex:
+		defaults = pionexDefaultEndpoints
 	case ProviderPoloniex:
 		defaults = poloniexDefaultEndpoints
 	case ProviderPyth:
 		defaults = pythDefaultEndpoints
+	case ProviderShade:
+		defaults = shadeDefaultEndpoints
 	case ProviderUniswapV3:
 		defaults = uniswapv3DefaultEndpoints
+	case ProviderUnstake:
+		defaults = unstakeDefaultEndpoints
+	case ProviderVelodromeV2:
+		defaults = velodromev2DefaultEndpoints
+	case ProviderWhitewhaleCmdx:
+		defaults = whitewhaleCmdxDefaultEndpoints
+	case ProviderWhitewhaleHuahua:
+		defaults = whitewhaleHuahuaDefaultEndpoints
+	case ProviderWhitewhaleInj:
+		defaults = whitewhaleInjDefaultEndpoints
+	case ProviderWhitewhaleJuno:
+		defaults = whitewhaleJunoDefaultEndpoints
+	case ProviderWhitewhaleLunc:
+		defaults = whitewhaleLuncDefaultEndpoints
+	case ProviderWhitewhaleLuna:
+		defaults = whitewhaleLunaDefaultEndpoints
+	case ProviderWhitewhaleSei:
+		defaults = whitewhaleSeiDefaultEndpoints
+	case ProviderWhitewhaleWhale:
+		defaults = whitewhaleWhaleDefaultEndpoints
 	case ProviderXt:
 		defaults = xtDefaultEndpoints
 	case ProviderZero:
@@ -403,6 +786,14 @@ func (e *Endpoint) SetDefaults() {
 			continue
 		}
 		e.ContractAddresses[symbol] = address
+	}
+
+	if e.VolumeBlocks == 0 {
+		e.VolumeBlocks = defaults.VolumeBlocks
+	}
+
+	if e.VolumePause <= 0 {
+		e.VolumePause = defaults.VolumePause
 	}
 }
 
@@ -470,11 +861,16 @@ func (p *provider) setPairs(
 	return nil
 }
 
-func (p *provider) setTickerPrice(symbol string, price sdk.Dec, volume sdk.Dec, timestamp time.Time) {
-	if price.IsZero() {
+func (p *provider) setTickerPrice(
+	symbol string,
+	price sdk.Dec,
+	volume sdk.Dec,
+	timestamp time.Time,
+) {
+	if price.IsNil() || price.LTE(sdk.ZeroDec()) {
 		p.logger.Warn().
 			Str("symbol", symbol).
-			Msg("price is zero")
+			Msgf("price is %s", price)
 		return
 	}
 
@@ -556,10 +952,28 @@ func (p *provider) getAllPairs() map[string]types.CurrencyPair {
 
 func (p *provider) getAvailablePairsFromContracts() (map[string]struct{}, error) {
 	symbols := map[string]struct{}{}
-	for symbol := range p.contracts {
+	for symbol := range p.endpoints.ContractAddresses {
 		symbols[symbol] = struct{}{}
 	}
 	return symbols, nil
+}
+
+func (p *provider) getPair(symbol string) (types.CurrencyPair, bool) {
+	pair, found := p.pairs[symbol]
+	if found {
+		return pair, true
+	}
+
+	pair, found = p.inverse[symbol]
+	if found {
+		return pair.Swap(), true
+	}
+
+	p.logger.Debug().
+		Str("symbol", symbol).
+		Msg("pair not found")
+
+	return pair, false
 }
 
 func (p *provider) getContractAddress(pair types.CurrencyPair) (string, error) {
@@ -580,6 +994,309 @@ func (p *provider) getContractAddress(pair types.CurrencyPair) (string, error) {
 		Err(err)
 
 	return "", err
+}
+
+// EVM specific
+
+func (p *provider) doEthCall(address, data string) (string, error) {
+	type Body struct {
+		Jsonrpc string        `json:"jsonrpc"`
+		Method  string        `json:"method"`
+		Params  []interface{} `json:"params"`
+		Id      int64         `json:"id"`
+	}
+
+	type Transaction struct {
+		To   string `json:"to"`
+		Data string `json:"data"`
+	}
+
+	type Response struct {
+		Result string `json:"result"` // Encoded data ex.: 0x0000000000000...
+	}
+
+	data = "0x" + strings.ReplaceAll(data, "0x", "")
+
+	body := Body{
+		Jsonrpc: "2.0",
+		Method:  "eth_call",
+		Params: []interface{}{
+			Transaction{
+				To:   address,
+				Data: data,
+			},
+			"latest",
+		},
+		Id: 1,
+	}
+
+	bz, err := json.Marshal(body)
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return "", err
+	}
+
+	content, err := p.httpPost("", bz)
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return "", err
+	}
+
+	var response Response
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		p.logger.Err(err).Msg("")
+		return "", err
+	}
+
+	return response.Result, nil
+}
+
+func (p *provider) getEthDecimals(contract string) (uint64, error) {
+	p.logger.Info().Str("contract", contract).Msg("get decimals")
+
+	hash, err := keccak256("decimals()")
+	if err != nil {
+		return 0, err
+	}
+
+	data := fmt.Sprintf("%0.8s%064d", hash, 0)
+
+	result, err := p.doEthCall(contract, data)
+	if err != nil {
+		return 0, err
+	}
+
+	types := []string{"uint8"}
+
+	decoded, err := decodeEthData(result, types)
+	if err != nil {
+		return 0, err
+	}
+
+	decimals, err := strconv.ParseUint(fmt.Sprintf("%v", decoded[0]), 10, 8)
+	if err != nil {
+		return 0, err
+	}
+
+	return decimals, nil
+}
+
+func (p *provider) evmRpcQuery(method, params string) (json.RawMessage, error) {
+	p.logger.Info().Str("method", method).Msg("query evm rpc")
+
+	query := []byte(fmt.Sprintf(
+		`{"jsonrpc":"2.0","method":"%s","params":[%s],"id":1}`,
+		method, string(params),
+	))
+
+	p.logger.Debug().Msg(string(query))
+
+	TelemetryEvmMethod(p.chain, p.name, method)
+
+	content, err := p.httpPost("", query)
+	if err != nil {
+		return nil, p.error(err)
+	}
+
+	var response struct {
+		Result json.RawMessage `json:"result"`
+	}
+
+	err = json.Unmarshal(content, &response)
+	if err != nil {
+		return nil, p.error(err)
+	}
+
+	if len(response.Result) == 0 {
+		p.logger.Error().
+			Str("content", string(content)).
+			Msg("empty return data")
+	}
+
+	return response.Result, nil
+}
+
+func (p *provider) getEvmHeight() (uint64, error) {
+	p.logger.Info().Msg("get height")
+
+	result, err := p.evmRpcQuery("eth_blockNumber", "")
+	if err != nil {
+		return 0, err
+	}
+
+	var height string
+	err = json.Unmarshal(result, &height)
+	if err != nil {
+		return 0, err
+	}
+	// strip "0x" from height string
+	return strconv.ParseUint(height[2:], 16, 64)
+}
+
+func (p *provider) evmGetLogs(
+	from, to uint64,
+	addresses, topics []string,
+) ([]EvmLog, error) {
+	// Note: getLogs returns values between and includin from- and end block
+	// -> from=1, to=3 returns [1, 2, 3]
+
+	type Params struct {
+		FromBlock string   `json:"fromBlock"`
+		ToBlock   string   `json:"toBlock"`
+		Address   []string `json:"address"`
+		Topics    []string `json:"topics"`
+	}
+
+	params := Params{
+		FromBlock: fmt.Sprintf("0x%x", from),
+		ToBlock:   fmt.Sprintf("0x%x", to),
+		Address:   addresses,
+		Topics:    topics,
+	}
+
+	if to == 0 {
+		params.ToBlock = "latest"
+	}
+
+	bz, err := json.Marshal(params)
+	if err != nil {
+		return nil, err
+	}
+
+	response, err := p.evmRpcQuery("eth_getLogs", string(bz))
+	if err != nil {
+		return nil, err
+	}
+
+	var logs []EvmLog
+	err = json.Unmarshal(response, &logs)
+	if err != nil {
+		p.logger.Err(err).Msg("failed to unmarshal evm log")
+		return nil, err
+	}
+
+	for i, log := range logs {
+		height, err := hexToUint64(log.Number)
+		if err != nil {
+			return nil, err
+		}
+		logs[i].Height = height
+	}
+
+	return logs, nil
+}
+
+func (p *provider) evmGetBlockByNumber(height uint64) (EvmBlock, error) {
+	params := fmt.Sprintf(`"0x%x",false`, height)
+	output, err := p.evmRpcQuery("eth_getBlockByNumber", params)
+	if err != nil {
+		return EvmBlock{}, err
+	}
+
+	var block EvmBlock
+
+	err = json.Unmarshal(output, &block)
+	if err != nil {
+		return EvmBlock{}, err
+	}
+
+	return block, nil
+}
+
+func (p *provider) evmCall(
+	address, method string,
+	args []string,
+) (json.RawMessage, error) {
+	p.logger.Info().Str("method", method).Msg("evmCall")
+
+	hash, err := keccak256(method)
+	if err != nil {
+		return nil, p.error(err)
+	}
+
+	data := "0x" + hash[:8]
+
+	for _, arg := range args {
+		data += arg
+	}
+
+	if len(args) == 0 {
+		// append 64 zeros
+		data = fmt.Sprintf("%s%064d", data, 0)
+	}
+
+	params := fmt.Sprintf(`{"to":"%s","data":"%s"},"latest"`, address, data)
+
+	return p.evmRpcQuery("eth_call", params)
+}
+
+func (p *provider) error(err error) error {
+	p.logger.Err(err).Msg("")
+	return err
+}
+
+func (p *provider) errorf(message string) error {
+	err := fmt.Errorf(message)
+	return p.error(err)
+}
+
+func hexToUint64(s string) (uint64, error) {
+	return strconv.ParseUint(strings.Replace(s, "0x", "", -1), 16, 64)
+}
+
+func decodeSqrtPrice(sqrt string) (sdk.Dec, error) {
+	dec := strToDec(sqrt)
+	if dec.IsZero() {
+		return sdk.Dec{}, fmt.Errorf("sqrt is 0")
+	}
+	return dec.Power(2).Quo(uintToDec(2).Power(192)), nil
+}
+
+func decodeEthData(data string, types []string) ([]interface{}, error) {
+	type AbiOutput struct {
+		Name         string `json:"name"`
+		Type         string `json:"type"`
+		InternalType string `json:"internalType"`
+	}
+
+	type AbiDefinition struct {
+		Name    string      `json:"name"`
+		Type    string      `json:"type"`
+		Outputs []AbiOutput `json:"outputs"`
+	}
+
+	outputs := []AbiOutput{}
+	for _, t := range types {
+		outputs = append(outputs, AbiOutput{
+			Name:         "",
+			Type:         t,
+			InternalType: t,
+		})
+	}
+
+	definition, err := json.Marshal([]AbiDefinition{{
+		Name:    "fn",
+		Type:    "function",
+		Outputs: outputs,
+	}})
+	if err != nil {
+		return nil, err
+	}
+
+	abi, err := abi.JSON(strings.NewReader(string(definition)))
+	if err != nil {
+		return nil, err
+	}
+
+	data = strings.TrimPrefix(data, "0x")
+
+	decoded, err := hex.DecodeString(data)
+	if err != nil {
+		return nil, err
+	}
+
+	return abi.Unpack("fn", decoded)
 }
 
 // String cast provider name to string.
@@ -623,6 +1340,10 @@ func checkHTTPStatus(resp *http.Response) error {
 }
 
 func strToDec(str string) sdk.Dec {
+	if len(str) == 0 {
+		return sdk.Dec{}
+	}
+
 	if strings.Contains(str, ".") {
 		split := strings.Split(str, ".")
 		if len(split[1]) > 18 {
@@ -630,11 +1351,24 @@ func strToDec(str string) sdk.Dec {
 			str = split[0] + "." + split[1][0:18]
 		}
 	}
-	return sdk.MustNewDecFromStr(str)
+	dec, err := sdk.NewDecFromStr(str)
+	if err != nil {
+		dec = sdk.Dec{}
+	}
+
+	return dec
+}
+
+func int64ToDec(i int64) sdk.Dec {
+	return strToDec(strconv.FormatInt(i, 10))
+}
+
+func uintToDec(u uint64) sdk.Dec {
+	return strToDec(strconv.FormatUint(u, 10))
 }
 
 func floatToDec(f float64) sdk.Dec {
-	return sdk.MustNewDecFromStr(strconv.FormatFloat(f, 'f', -1, 64))
+	return strToDec(strconv.FormatFloat(f, 'f', -1, 64))
 }
 
 func invertDec(d sdk.Dec) sdk.Dec {
@@ -642,4 +1376,51 @@ func invertDec(d sdk.Dec) sdk.Dec {
 		return sdk.ZeroDec()
 	}
 	return sdk.NewDec(1).Quo(d)
+}
+
+func computeDecimalsFactor(base, quote int64) (sdk.Dec, error) {
+	delta := base - quote
+	factor := uintToDec(1)
+	if delta == 0 {
+		return factor, nil
+	} else if delta < 0 {
+		factor = factor.Quo(uintToDec(10).Power(uint64(delta * -1)))
+	} else {
+		factor = uintToDec(10).Power(uint64(delta))
+	}
+
+	return factor, nil
+}
+
+func keccak256(s string) (string, error) {
+	hash := sha3.NewLegacyKeccak256()
+	_, err := hash.Write([]byte(s))
+	if err != nil {
+		return "", err
+	}
+	return fmt.Sprintf("%x", hash.Sum(nil)), nil
+}
+
+func parseDenom(s string) (sdk.Dec, string, error) {
+	re := regexp.MustCompile(`^([0-9]+)(.*)$`)
+
+	matches := re.FindAllStringSubmatch(s, -1)[0]
+	if len(matches) != 3 {
+		return sdk.Dec{}, "", fmt.Errorf("failed parsing denom")
+	}
+
+	amount := strToDec(matches[1])
+
+	return amount, matches[2], nil
+}
+
+func (b *EvmBlock) GetTime() (time.Time, error) {
+	seconds, err := strconv.ParseInt(
+		strings.Replace(b.Timestamp, "0x", "", -1), 16, 64,
+	)
+	if err != nil {
+		return time.Time{}, err
+	}
+
+	return time.Unix(seconds, 0), nil
 }
