@@ -3,6 +3,7 @@ package oracle
 import (
 	"context"
 	"crypto/rand"
+	"database/sql"
 	"encoding/hex"
 	"fmt"
 	"math"
@@ -37,6 +38,11 @@ import (
 const (
 	tickerSleep = 1000 * time.Millisecond
 )
+
+type ProviderWeight struct {
+	Type   string
+	Weight map[string]sdk.Dec
+}
 
 // PreviousPrevote defines a structure for defining the previous prevote
 // submitted on-chain.
@@ -75,6 +81,10 @@ type Oracle struct {
 	derivativePairs      map[string][]types.CurrencyPair
 	derivativeSymbols    map[string]struct{}
 	contractAddresses    map[string]map[string]string
+	providerWeights      map[string]ProviderWeight
+	decimals             map[string]map[string]int
+	periods              map[string]map[string]int
+	volumeDatabase       *sql.DB
 
 	mtx             sync.RWMutex
 	lastPriceSyncTS time.Time
@@ -97,6 +107,10 @@ func New(
 	healthchecksConfig []config.Healthchecks,
 	history history.PriceHistory,
 	contractAddresses map[string]map[string]string,
+	providerWeights map[string]ProviderWeight,
+	decimals map[string]map[string]int,
+	periods map[string]map[string]int,
+	volumeDatabase *sql.DB,
 ) *Oracle {
 	providerPairs := make(map[provider.Name][]types.CurrencyPair)
 	for _, pair := range currencyPairs {
@@ -120,6 +134,7 @@ func New(
 			}
 		}
 	}
+
 	return &Oracle{
 		logger:               logger.With().Str("module", "oracle").Logger(),
 		closer:               pfsync.NewCloser(),
@@ -138,6 +153,10 @@ func New(
 		derivativeSymbols:    derivativeDenoms,
 		history:              history,
 		contractAddresses:    contractAddresses,
+		providerWeights:      providerWeights,
+		decimals:             decimals,
+		periods:              periods,
+		volumeDatabase:       volumeDatabase,
 	}
 }
 
@@ -216,10 +235,15 @@ func (o *Oracle) SetPrices(ctx context.Context) error {
 		priceProvider, found := o.priceProviders[providerName]
 		if !found {
 			endpoint := o.endpoints[providerName]
-			contractAddresses, _ := o.contractAddresses[providerName.String()]
+			contractAddresses := o.contractAddresses[providerName.String()]
+			decimals := o.decimals[providerName.String()]
+			periods := o.periods[providerName.String()]
 			endpoint.ContractAddresses = contractAddresses
+			endpoint.Decimals = decimals
+			endpoint.Periods = periods
 
 			newProvider, err := NewProvider(
+				o.volumeDatabase,
 				ctx,
 				providerName,
 				o.logger,
@@ -343,6 +367,7 @@ func (o *Oracle) SetPrices(ctx context.Context) error {
 		o.providerPairs,
 		o.deviations,
 		o.providerMinOverrides,
+		o.providerWeights,
 	)
 	if err != nil {
 		return err
@@ -377,6 +402,7 @@ func GetComputedPrices(
 	providerPairs map[provider.Name][]types.CurrencyPair,
 	deviations map[string]sdk.Dec,
 	providerMinOverrides map[string]int,
+	providerWeights map[string]ProviderWeight,
 ) (prices map[string]sdk.Dec, err error) {
 	rates, err := convertTickersToUSD(
 		logger,
@@ -384,6 +410,7 @@ func GetComputedPrices(
 		providerPairs,
 		deviations,
 		providerMinOverrides,
+		providerWeights,
 	)
 	if err != nil {
 		return nil, err
@@ -436,6 +463,7 @@ func (o *Oracle) GetParams(ctx context.Context) (oracletypes.Params, error) {
 }
 
 func NewProvider(
+	db *sql.DB,
 	ctx context.Context,
 	providerName provider.Name,
 	logger zerolog.Logger,
@@ -446,10 +474,17 @@ func NewProvider(
 	providerLogger := logger.With().Str("provider", providerName.String()).Logger()
 	switch providerName {
 
-	case provider.ProviderAstroportNeutron, provider.ProviderAstroportTerra2, provider.ProviderAstroportInjective:
+	case
+		provider.ProviderAstroportNeutron,
+		provider.ProviderAstroportTerra2,
+		provider.ProviderAstroportInjective:
 		return provider.NewAstroportProvider(ctx, providerLogger, endpoint, providerPairs...)
-	case provider.ProviderBinance, provider.ProviderBinanceUS:
+	case
+		provider.ProviderBinance,
+		provider.ProviderBinanceUS:
 		return provider.NewBinanceProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderBingx:
+		return provider.NewBingxProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderBitfinex:
 		return provider.NewBitfinexProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderBitget:
@@ -463,19 +498,25 @@ func NewProvider(
 	case provider.ProviderBybit:
 		return provider.NewBybitProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderCamelotV2, provider.ProviderCamelotV3:
-		return provider.NewCamelotProvider(ctx, providerLogger, endpoint, providerPairs...)
+		return provider.NewCamelotProvider(db, ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderCoinbase:
 		return provider.NewCoinbaseProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderCoinex:
+		return provider.NewCoinexProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderCrypto:
 		return provider.NewCryptoProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderCurve:
 		return provider.NewCurveProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderDexter:
+		return provider.NewDexterProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderFin:
 		return provider.NewFinProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderFinV2:
-		return provider.NewFinV2Provider(ctx, providerLogger, endpoint, providerPairs...)
+		return provider.NewFinV2Provider(db, ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderGate:
 		return provider.NewGateProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderHelix:
+		return provider.NewHelixProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderHitBtc:
 		return provider.NewHitBtcProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderHuobi:
@@ -488,6 +529,8 @@ func NewProvider(
 		return provider.NewKucoinProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderLbank:
 		return provider.NewLbankProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderMaya:
+		return provider.NewMayaProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderMexc:
 		return provider.NewMexcProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderMock:
@@ -497,15 +540,35 @@ func NewProvider(
 	case provider.ProviderOsmosis:
 		return provider.NewOsmosisProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderOsmosisV2:
-		return provider.NewOsmosisV2Provider(ctx, providerLogger, endpoint, providerPairs...)
+		return provider.NewOsmosisV2Provider(db, ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderPancakeV3Bsc:
+		return provider.NewPancakeProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderPhemex:
 		return provider.NewPhemexProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderPionex:
+		return provider.NewPionexProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderPoloniex:
 		return provider.NewPoloniexProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderPyth:
 		return provider.NewPythProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderShade:
+		return provider.NewShadeProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderUniswapV3:
 		return provider.NewUniswapV3Provider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderUnstake:
+		return provider.NewUnstakeProvider(ctx, providerLogger, endpoint, providerPairs...)
+	case provider.ProviderVelodromeV2:
+		return provider.NewVelodromeV2Provider(ctx, providerLogger, endpoint, providerPairs...)
+	case
+		provider.ProviderWhitewhaleCmdx,
+		provider.ProviderWhitewhaleHuahua,
+		provider.ProviderWhitewhaleInj,
+		provider.ProviderWhitewhaleJuno,
+		provider.ProviderWhitewhaleLunc,
+		provider.ProviderWhitewhaleLuna,
+		provider.ProviderWhitewhaleSei,
+		provider.ProviderWhitewhaleWhale:
+		return provider.NewWhitewhaleProvider(db, ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderXt:
 		return provider.NewXtProvider(ctx, providerLogger, endpoint, providerPairs...)
 	case provider.ProviderZero:
@@ -525,7 +588,12 @@ func (o *Oracle) checkWhitelist(params oracletypes.Params) {
 }
 
 func (o *Oracle) tick(ctx context.Context) error {
-	o.logger.Debug().Msg("executing oracle tick")
+	o.logger.Info().Msg("executing oracle tick")
+
+	// Create and start all provider routines immediately
+	if len(o.priceProviders) == 0 {
+		o.SetPrices(ctx)
+	}
 
 	blockHeight, err := o.oracleClient.ChainHeight.GetChainHeight()
 	if err != nil {
